Move some libraries to src/utils
Verified that demo.conf still works.
